refactor: move get-on-chain-txs to its own use-case
