+++ src/app/wallets/get-transactions-by-addresses.ts
@@ -1,19 +1,16 @@
-import { BTC_NETWORK, ONCHAIN_MIN_CONFIRMATIONS, SECS_PER_10_MINS } from "@config"
+import { ONCHAIN_MIN_CONFIRMATIONS } from "@config"
 
 import { getCurrentPrice } from "@app/prices"
 import { PartialResult } from "@app/partial-result"
 
-import { CacheKeys } from "@domain/cache"
 import { LedgerError } from "@domain/ledger"
 import { WalletTransactionHistory } from "@domain/wallets"
-import { OnChainError, TxDecoder, TxFilter } from "@domain/bitcoin/onchain"
+import { TxFilter } from "@domain/bitcoin/onchain"
 
 import { baseLogger } from "@services/logger"
 import { LedgerService } from "@services/ledger"
-import { RedisCacheService } from "@services/cache"
-import { OnChainService } from "@services/lnd/onchain-service"
 
-const redisCache = RedisCacheService()
+import getOnChainTxs from "./get-on-chain-txs"
 
 export const getTransactionsForWalletsByAddresses = async ({
   wallets,
@@ -34,22 +31,7 @@ export const getTransactionsForWalletsByAddresses = async ({
 
   const confirmedHistory = WalletTransactionHistory.fromLedger(ledgerTransactions)
 
-  const listIncomingToUpdateCache = async () => {
-    const onChain = OnChainService(TxDecoder(BTC_NETWORK))
-    if (onChain instanceof OnChainError) {
-      baseLogger.warn({ onChain }, "impossible to create OnChainService")
-      return onChain
-    }
-    return onChain.listIncomingTransactions(ONCHAIN_MIN_CONFIRMATIONS)
-  }
-
-  // we are getting both the transactions in the mempool and the transaction that
-  // have been mined by not yet credited because they haven't reached enough confirmations
-  const onChainTxs = await redisCache.getOrSet({
-    key: CacheKeys.LastOnChainTransactions,
-    ttlSecs: SECS_PER_10_MINS,
-    fn: listIncomingToUpdateCache,
-  })
+  const onChainTxs = await getOnChainTxs()
   if (onChainTxs instanceof Error) {
     baseLogger.warn({ onChainTxs }, "impossible to get listIncomingTransactions")
     return PartialResult.partial(confirmedHistory.transactions, onChainTxs)
