+++ src/app/wallets/get-transactions-for-wallet.ts
@@ -1,21 +1,18 @@
-import { BTC_NETWORK, ONCHAIN_MIN_CONFIRMATIONS, SECS_PER_10_MINS } from "@config"
+import { ONCHAIN_MIN_CONFIRMATIONS } from "@config"
 
 import { getCurrentPrice } from "@app/prices"
 import { PartialResult } from "@app/partial-result"
 
-import { CacheKeys } from "@domain/cache"
 import { LedgerError } from "@domain/ledger"
 import { RepositoryError } from "@domain/errors"
 import { WalletTransactionHistory } from "@domain/wallets"
-import { OnChainError, TxDecoder, TxFilter } from "@domain/bitcoin/onchain"
+import { TxFilter } from "@domain/bitcoin/onchain"
 
 import { baseLogger } from "@services/logger"
 import { LedgerService } from "@services/ledger"
-import { RedisCacheService } from "@services/cache"
 import { WalletsRepository } from "@services/mongoose"
-import { OnChainService } from "@services/lnd/onchain-service"
 
-const redisCache = RedisCacheService()
+import getOnChainTxs from "./get-on-chain-txs"
 
 // FIXME(nicolas): remove only used in tests
 export const getTransactionsForWalletId = async ({
@@ -41,22 +38,7 @@ export const getTransactionsForWallets = async (
 
   const confirmedHistory = WalletTransactionHistory.fromLedger(ledgerTransactions)
 
-  const listIncomingToUpdateCache = async () => {
-    const onChain = OnChainService(TxDecoder(BTC_NETWORK))
-    if (onChain instanceof OnChainError) {
-      baseLogger.warn({ onChain }, "impossible to create OnChainService")
-      return onChain
-    }
-    return onChain.listIncomingTransactions(ONCHAIN_MIN_CONFIRMATIONS)
-  }
-
-  // we are getting both the transactions in the mempool and the transaction that
-  // have been mined by not yet credited because they haven't reached enough confirmations
-  const onChainTxs = await redisCache.getOrSet({
-    key: CacheKeys.LastOnChainTransactions,
-    ttlSecs: SECS_PER_10_MINS,
-    fn: listIncomingToUpdateCache,
-  })
+  const onChainTxs = await getOnChainTxs()
   if (onChainTxs instanceof Error) {
     baseLogger.warn({ onChainTxs }, "impossible to get listIncomingTransactions")
     return PartialResult.partial(confirmedHistory.transactions, onChainTxs)
